requirements.txt
py-cpuinfo
brian2
